refactor: move more api stuff around
