+++ src/actions/submit/submit_prs.ts
@@ -2,7 +2,7 @@ import graphiteCLIRoutes from '@withgraphite/graphite-cli-routes';
 import * as t from '@withgraphite/retype';
 import { request } from '@withgraphite/retyped-routes';
 import chalk from 'chalk';
-import { API_SERVER } from '../../lib/api';
+import { API_SERVER } from '../../lib/api/server';
 import { TContext } from '../../lib/context';
 import { ExitFailedError, PreconditionsFailedError } from '../../lib/errors';
 import { cuteString } from '../../lib/utils/cute_string';
