+++ src/lib/telemetry/tracer.ts
@@ -2,7 +2,7 @@
 import graphiteCLIRoutes from '@withgraphite/graphite-cli-routes';
 import { request } from '@withgraphite/retyped-routes';
 import { version } from '../../../package.json';
-import { API_SERVER } from '../api';
+import { API_SERVER } from '../api/server';
 import { cuteString } from '../utils/cute_string';
 
 type spanNameT = 'function' | 'execSync' | 'command';
