+++ test/fast/commands/repo/sync.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { execSync } from 'child_process';
 import fs from 'fs-extra';
 import nock from 'nock';
-import { API_SERVER } from '../../../../src/lib/api';
+import { API_SERVER } from '../../../../src/lib/api/server';
 import { allScenes } from '../../../lib/scenes/all_scenes';
 import { configureTest } from '../../../lib/utils/configure_test';
 import { expectBranches } from '../../../lib/utils/expect_branches';
